improved code
